feat: 레디스 설정 값 및 유저 기록값 타입 통일

+++ dailyquest_web/src/main/java/dailyquest/admin/dto/SystemSettingsRequest.kt
@@ -1,7 +1,7 @@
 package dailyquest.admin.dto
 
 class SystemSettingsRequest(
-    val questClearExp: Int,
-    val questClearGold: Int,
-    val maxRewardCount: Int,
+    val questClearExp: Long,
+    val questClearGold: Long,
+    val maxRewardCount: Long,
 )
